@@ -1,6 +1,6 @@
 package com.mysticalducks.rest.finance.exception;
 
-import javax.persistence.EntityNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 public class EmptyBodyException extends EntityNotFoundException {
 	private static final long serialVersionUID = -9151144634415261762L;
