@@ -8,19 +8,18 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import javax.servlet.http.HttpServletResponse;
-
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.MissingServletRequestParameterException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import jakarta.servlet.http.HttpServletResponse;
 
 //@EnableWebMvc
 @ControllerAdvice
@@ -86,18 +85,18 @@ public class ExceptionHandlerAdvice extends ResponseEntityExceptionHandler {
 	}
 
 	@Override
-	protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex,
-			HttpHeaders headers, HttpStatus status, WebRequest request) {
+	protected ResponseEntity<Object> handleMethodArgumentNotValid( MethodArgumentNotValidException ex, HttpHeaders headers, HttpStatusCode status, WebRequest request) {
 
-		Map<String, Object> body = new LinkedHashMap<>();
-		body.put("timestamp", LocalDate.now());
-		body.put("status", status.value());
+	    Map<String, Object> body = new LinkedHashMap<>();
+	    body.put("timestamp", LocalDate.now());
+	    body.put("status", status.value());
 
-		List<String> errors = ex.getBindingResult().getFieldErrors().stream().map(x -> x.getDefaultMessage())
-				.collect(Collectors.toList());
+	    List<String> errors = ex.getBindingResult().getFieldErrors().stream()
+	            .map(x -> x.getDefaultMessage())
+	            .collect(Collectors.toList());
 
-		body.put("errors", errors);
+	    body.put("errors", errors);
 
-		return new ResponseEntity<>(body, HttpStatus.BAD_REQUEST);
+	    return new ResponseEntity<>(body, HttpStatus.BAD_REQUEST);
 	}
 }
