@@ -26,10 +26,11 @@ public class FinanceApplication {
 	public WebMvcConfigurer corsConfigurer() {
 		return new WebMvcConfigurer() {
 			public void addCorsMappings(CorsRegistry registry) {
-				registry.addMapping("/**").allowedOrigins("http://localhost:9000");
+				registry.addMapping("/**").allowedOrigins("http://localhost:9000", "http://localhost:4200");
 			}
 		};
 	}
+	
 	@Bean
     public OpenAPI springShopOpenAPI() {
         return new OpenAPI()
