@@ -1,7 +1,9 @@
 package com.mysticalducks.rest.finance.exception;
 
 public class ChatNotFoundException extends RuntimeException { 
-    public ChatNotFoundException(String iconMessage) {
+	private static final long serialVersionUID = -8068031606737125601L;
+
+	public ChatNotFoundException(String iconMessage) {
         super(iconMessage);
     }
     
