@@ -4,9 +4,9 @@ import java.util.Date;
 
 public class ErrorDetails {
 	
-	private Date timestamp;
-	private String message;
-	private String details;
+	Date timestamp;
+	String message;
+	String details;
 	
 	public ErrorDetails(Date timestamp, String message) {
 		this.timestamp = timestamp;
